+++ rtc_base/rate_statistics.h
@@ -42,7 +42,7 @@ class RTC_EXPORT RateStatistics {
   void Reset();
 
   // Update rate with a new data point, moving averaging window as needed.
-  void Update(size_t count, int64_t now_ms);
+  void Update(uint64_t count, int64_t now_ms);
 
   // Note that despite this being a const method, it still updates the internal
   // state (moves averaging window), but it doesn't make any alterations that
@@ -50,7 +50,7 @@ class RTC_EXPORT RateStatistics {
   // from a monotonic clock. Ie, it doesn't matter if this call moves the
   // window, since any subsequent call to Update or Rate would still have moved
   // the window as much or more.
-  absl::optional<uint32_t> Rate(int64_t now_ms) const;
+  absl::optional<uint64_t> Rate(int64_t now_ms) const;
 
   // Update the size of the averaging window. The maximum allowed value for
   // window_size_ms is max_window_size_ms as supplied in the constructor.
@@ -63,13 +63,13 @@ class RTC_EXPORT RateStatistics {
   // Counters are kept in buckets (circular buffer), with one bucket
   // per millisecond.
   struct Bucket {
-    size_t sum;      // Sum of all samples in this bucket.
+    uint64_t sum;    // Sum of all samples in this bucket.
     size_t samples;  // Number of samples in this bucket.
   };
   std::unique_ptr<Bucket[]> buckets_;
 
   // Total count recorded in buckets.
-  size_t accumulated_count_;
+  uint64_t accumulated_count_;
 
   // The total number of samples in the buckets.
   size_t num_samples_;
